Move CLI related code to new cli module
Simplify unit tests (pytest style)
