+++ tox.ini
@@ -20,7 +20,7 @@ commands = flake8
 deps =
     pylint
     -rrequirements.txt
-commands = pylint carddav2hatchbuck setup
+commands = pylint carddav2hatchbuck setup tests
 
 [flake8]
 exclude = .tox,carddav,.vdirsyncer,venv,*.egg*
